+++ src/basic/arphrd-list.c
@@ -25,7 +25,7 @@
 #include "util.h"
 #include "arphrd-list.h"
 
-static const struct arphrd_name* lookup_arphrd(register const char *str, register unsigned int len);
+static const struct arphrd_name* lookup_arphrd(register const char *str, register size_t len);
 
 #include "arphrd-to-name.h"
 #include "arphrd-from-name.h"
